requirements.txt
flask
flask-cors
mysql-connector-python
werkzeug
bcrypt
psycopg2-binary
